Added a proper queue and web app
This uses a nicer queue for new messages. It will expire them after 30
seconds, so even if Photo Booth shits the bed, the queue won't get too
terribly out of sync.

This also does a nicer web app for the remote. It's got a proper icon
and the metadata to open it full screen and all. It sizes appropriately
for different devices, and reacts better when the keyboard is opened.

Add the page to your home screen to get the app behaviour.